Remove nulls, use explicit return
The iterate function uses a nested for loop.  Without an explicit
return, the for loop acts like a comprehension and the whole list of
results it cached and returned.  The resulting JavaScript is ugly.  By
adding an explicit return, the CoffeeScript compiler compiles it to a
normal for loop.

Previously I was using null to force the return value.  But I like using
an explicit return better (mainly because when I looked at the code
months later, I had no idea why I had a bunch of functions that ended
with null.)